Use tag date for sugestions
This branch uses the tag date to check for newer dates and make
sugestions for updates. This is safer than trying to parse
random version numbers, but still requires manual revision
because older versions can be still under maintenance. Anyway,
the number of tags for older versions that are more recent than
the current one is usually very small (one or two), so it isn't
really a problem.